requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
asyncpg==0.29.0
pgvector==0.2.4
google-generativeai==0.3.2
firebase-admin==6.4.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0